initial upload
The keyboards of many Series-80 computers suffer from splitting plungers.

This model can bes used to print a replacement plunger for one of the several variations of these pungers.

In contrast to other models floating around, the reinforcement on the top edge is broken in the moddle of each edge.
This allows pressing the key cap down into the plunger to the same height as the original part.
Thus the key is level with its neighboring keys.

The required low wall thickness and smooth surface finish require that this part is printed with a resin printer - a normal additive printer will not produce useable results.
The resin parts are more brittle than e.g. PLA or ABS parts, but strong enough for normal usage.
A thin application of Teflon spray to the gliding surface can also be helpful to ensure smooth operation.
I also increase the spring length slightly (by about 5 mm) to increase the spring force slightly.

Don't forget to add the spring before mounting the plunger and insert two narrow strips of paper to guide the center bar between the spring contacts.

I have used it to replace a broken plunger on my HP-87.